VR: Add .register operation
We found two bugs in the VRR paper relating to the client table:

1. a correctness bug, where successive client crashes may cause request
numbers to collide for different request payloads, resulting in
requests receiving the wrong reply, and

2. a liveness bug, where if the client table is updated for `.request`
and `.prepare` messages with the client's latest request number, then
the client may be locked out from the cluster if the request is ever
reordered through a view change.

We therefore take a different approach with the implementation of our
client table, to:

1. register client sessions explicitly through the state machine to
ensure that client session numbers always increase, and

2. make a more careful distinction between uncommitted and committed
request numbers, considering that uncommitted requests may not survive
a view change.

This is the start of adding the `.register` operation to register a new
client session with the cluster.